requirements.txt
streamlit
PyMuPDF
python-docx
groq
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
optimum[onnxruntime]